requirements.txt
aiohttp
asyncpg
discord.py @ git+https://github.com/Rapptz/discord.py@master
ruamel.yaml
python-dotenv
stringcase
tweepy[async]
python-dateutil
typing-extensions
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
google-cloud-pubsub
emoji
